fix: subsequent renames are being handled properly
previously when a file which was already updated was renamed then it was no longer in cloudfiles and was thus considered to be a new file, fixed that by introducing "renamedWhileOffline".
also while uploading modified files which were renamed after the modify item was added to pending sync, the readFile would try to get the file by its previous name, solved this by introducing "finalNamesForFileID" which keeps track of the final names for each fileId which is used for every rename operation while online.
new uploads are assigned dummy ids to keep track of them in other operations.